# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy DB Service API
  description: User persistence API for the fantasy hockey tool.
  version: v1
servers:
- url: /
tags:
- name: User projections
  description: "Saved player projections, scoped to a user"
- name: Users
  description: User management
paths:
  /api/v1/users/{userId}/projections/{id}:
    get:
      tags:
      - User projections
      summary: "Fetch a single saved projection, including its data"
      operationId: get
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Projection found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    put:
      tags:
      - User projections
      summary: Update an existing projection
      operationId: update
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectionRequest"
        required: true
      responses:
        "200":
          description: Projection updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          description: Validation failed (blank name/data or too large)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "409":
          description: The user already has another projection with that name
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    delete:
      tags:
      - User projections
      summary: Delete a saved projection
      operationId: delete
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Projection deleted
        "404":
          description: No such projection for this user
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Look up a user by email address
      operationId: getByEmail
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: No user with that email
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    post:
      tags:
      - Users
      summary: Create a new user
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: User created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation failed (blank email / password hash)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: A user with that email already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/{userId}/projections:
    get:
      tags:
      - User projections
      summary: "List a user's saved projections (metadata only, without the projection\
        \ data)"
      operationId: list
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: "The user's projections, newest first"
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectionSummaryResponse"
    post:
      tags:
      - User projections
      summary: Save a new projection for a user
      operationId: create_1
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectionRequest"
        required: true
      responses:
        "201":
          description: Projection created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          description: Validation failed (blank name/data or too large)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "409":
          description: The user already has a projection with that name
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
  /api/v1/users/exists:
    get:
      tags:
      - Users
      summary: Check whether a user with the given email exists
      operationId: existsByEmail
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Always returns 200; check the 'exists' field
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExistsResponse"
components:
  schemas:
    UpdateProjectionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        data:
          type: string
          maxLength: 1000000
          minLength: 0
      required:
      - data
      - name
    ProjectionResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        data:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - data
      - id
      - name
      - updatedAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          minLength: 0
        passwordHash:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - email
      - passwordHash
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        passwordHash:
          type: string
      required:
      - email
      - id
      - passwordHash
    CreateProjectionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        data:
          type: string
          maxLength: 1000000
          minLength: 0
      required:
      - data
      - name
    ProjectionSummaryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - id
      - name
      - updatedAt
    ExistsResponse:
      type: object
      properties:
        exists:
          type: boolean
      required:
      - exists